Reorganize the namespaces of the project
