+++ GuildWars2.PvPCasterToolbox/Controls/ImageModulationConfig.xaml
@@ -3,7 +3,7 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:local="clr-namespace:GuildWars2.PvPCasterToolbox"
+        xmlns:local="clr-namespace:GuildWars2.PvPCasterToolbox.Controls"
         mc:Ignorable="d" 
         d:DesignHeight="250" d:DesignWidth="300"
         x:Name="Root">
